chore: refactor Exceptions into separate namespaces
