+++ src/OrasProject.Oras/Content/Digest.cs
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using OrasProject.Oras.Exceptions;
+using OrasProject.Oras.Content.Exceptions;
 using System;
 using System.Collections.Generic;
 using System.Security.Cryptography;
