+++ src/OrasProject.Oras/Registry/Exceptions/InvalidReferenceException.cs
@@ -13,8 +13,7 @@
 
 using System;
 
-namespace OrasProject.Oras.Exceptions;
-
+namespace OrasProject.Oras.Registry.Exceptions;
 /// <summary>
 /// InvalidReferenceException is thrown when the reference is invalid.
 /// </summary>
