+++ src/OrasProject.Oras/Registry/Reference.cs
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using OrasProject.Oras.Exceptions;
+using OrasProject.Oras.Registry.Exceptions;
 using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Text.RegularExpressions;
