+++ src/OrasProject.Oras/Registry/Remote/UriFactory.cs
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using OrasProject.Oras.Exceptions;
+using OrasProject.Oras.Registry.Exceptions;
 using System;
 using System.Web;
