+++ tests/OrasProject.Oras.Tests/Content/ExtensionsTest.cs
@@ -17,6 +17,7 @@ using System.Text;
 using System.Text.Json;
 using Moq;
 using OrasProject.Oras.Content;
+using OrasProject.Oras.Content.Exceptions;
 using static OrasProject.Oras.Tests.Remote.Util.RandomDataGenerator;
 using OrasProject.Oras.Oci;
 using OrasProject.Oras.Registry;
