+++ tests/OrasProject.Oras.Tests/Exceptions/ExceptionTest.cs
@@ -11,7 +11,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using OrasProject.Oras.Content.Exceptions;
 using OrasProject.Oras.Exceptions;
+using OrasProject.Oras.Registry.Exceptions;
+using OrasProject.Oras.Registry.Remote.Exceptions;
 using Xunit;
 
 namespace OrasProject.Oras.Tests.Exceptions;
