+++ tests/OrasProject.Oras.Tests/Remote/ReferenceTest.cs
@@ -11,8 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using OrasProject.Oras.Exceptions;
 using OrasProject.Oras.Registry;
+using OrasProject.Oras.Registry.Exceptions;
 using Xunit;
 
 namespace OrasProject.Oras.Tests.Remote;
