+++ tests/OrasProject.Oras.Tests/Remote/ReferrersTest.cs
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using OrasProject.Oras.Exceptions;
+using OrasProject.Oras.Content.Exceptions;
 using OrasProject.Oras.Oci;
 using OrasProject.Oras.Registry.Remote;
 using static OrasProject.Oras.Tests.Remote.Util.Util;
